// src/app/shared/components/dashboard-integrated-container/chart-types.ts
export enum ChartTypes {
    BAR_CHART = 'bar-chart',
    PIE_CHART = 'pie-chart',
    DONUT_CHART = 'donut-chart',
    BAR_LINE_CHART = 'bar-line-chart',
    AREA_CHART = 'area-chart',
    BUBBLE_CHART = 'bubble-chart',
    DATA_GRID = 'data-grid',
    SCATTER_CHART = 'scatter-chart',
    INDONESIA_MAP = 'indonesia-map',
    USA_MAP = 'usa-map',
    WORLD_MAP = 'world-map',
    PIVOT_TABLE = 'pivot-table',
    SCORE_CARD = 'score-card',
    NETWORK_GRAPH = 'network-graph',
    FORM_CARD = 'form-card',
    IMAGE = 'image',
    TIMELINE = 'time-line',
    HEAT_MAP_2 = 'heat-map2',
    HEAT_MAP = 'heat-map',
    SANKEY = 'sankey-chart',
    TREE_HEAT_MAP = 'tree-heat-map',
    WORLD_MAP_BUBBLE = 'world-map-bubble',
    STACKED_BAR_CHART = 'stacked-bar-chart',
    SCATTER_3D = 'scatter3d'
}